+++ src/phishbench_cli/main.py
@@ -8,6 +8,7 @@ from typing import List, Dict
 
 import joblib
 import pandas as pd
+from sklearn.model_selection import train_test_split
 
 import phishbench
 import phishbench.classification as classification
@@ -89,11 +90,9 @@ def extract_train_features(pickle_dir: str,
         out_path = os.path.join(pickle_dir, 'train_features.csv')
         export_features_to_csv(feature_list_dict_train, y_train, out_path)
 
-    # Tranform the list of dictionaries into a sparse matrix
+    # Transform the list of dictionaries into a sparse matrix
     vectorizer = preprocessing.Vectorizer()
     x_train = vectorizer.fit_transform(feature_list_dict_train)
-
-    joblib.dump(x_train, os.path.join(pickle_dir, "X_train_unprocessed.pkl"))
 
     return x_train, y_train, vectorizer
 
@@ -211,6 +210,8 @@ def extract_features(extraction_module: ModuleType):
         joblib.dump(x_test, os.path.join(pickle_dir, "X_test.pkl"))
         joblib.dump(y_test, os.path.join(pickle_dir, "y_test.pkl"))
         phishbench_globals.logger.info("Feature Extraction for testing dataset: Done!")
+    elif phishbench_globals.config["Extraction"].getboolean('split dataset'):
+        x_train, x_test, y_train, y_test = train_test_split(x_train, y_train)
     else:
         x_test = None
         y_test = None